@@ -1,65 +1,63 @@
-import "./ExpenseForm.css";
-import { useState } from "react"; //55
+import React, { useState } from "react";
 
-const ExpenseForm = () => {
-  const [enteredTitle, setEnteredTitle] = useState(""); // but i can go another way
+import "./ExpenseForm.css";
+
+const ExpenseForm = (props) => {
+  const [enteredTitle, setEnteredTitle] = useState("");
   const [enteredAmount, setEnteredAmount] = useState("");
   const [enteredDate, setEnteredDate] = useState("");
-
   // const [userInput, setUserInput] = useState({
-  //   enteredTitle: "",
-  //   enteredAmount: "",
-  //   enteredDate: "",
+  //   enteredTitle: '',
+  //   enteredAmount: '',
+  //   enteredDate: '',
   // });
 
   const titleChangeHandler = (event) => {
     setEnteredTitle(event.target.value);
-    //   setUserInput({
-    //     ...userInput,
-    //     enteredTitle: event.target.value,
+    // setUserInput({
+    //   ...userInput,
+    //   enteredTitle: event.target.value,
+    // });
+    // setUserInput((prevState) => {
+    //   return { ...prevState, enteredTitle: event.target.value };
     // });
   };
 
-  //   setUserState({ // do this when my state depends on previous state
-  //     return {..prevState, enteredTitle: event.target.value};
-  //   });
-  // };
-
   const amountChangeHandler = (event) => {
     setEnteredAmount(event.target.value);
-    //   setUserInput({
-    //     ...userInput,
-    //     enteredAmount: event.target.value,
-    //   });
+    // setUserInput({
+    //   ...userInput,
+    //   enteredAmount: event.target.value,
+    // });
   };
 
   const dateChangeHandler = (event) => {
     setEnteredDate(event.target.value);
-    //   setUserInput({
-    //     ...userInput,
-    //     enteredDate: event.target.value,
-    //   });
+    // setUserInput({
+    //   ...userInput,
+    //   enteredDate: event.target.value,
+    // });
   };
 
   const submitHandler = (event) => {
     event.preventDefault();
 
     const expenseData = {
-      // don't need to combine while i use alternative way of useState declaration
       title: enteredTitle,
       amount: enteredAmount,
       date: new Date(enteredDate),
     };
-    console.log(expenseData);
-    setEnteredTitle('');
-    setEnteredAmount('');
-    setEnteredDate('');
+
+    props.onSaveExpenseData(expenseData);
+    setEnteredTitle("");
+    setEnteredAmount("");
+    setEnteredDate("");
   };
 
   return (
     <form onSubmit={submitHandler}>
       <div className="new-expense__controls">
-        <div className="new-expense__controls">
+        <div className="new-expense__control">
           <label>Title</label>
           <input
             type="text"
@@ -67,7 +65,7 @@ const ExpenseForm = () => {
             onChange={titleChangeHandler}
           />
         </div>
-        <div className="new-expense__controls">
+        <div className="new-expense__control">
           <label>Amount</label>
           <input
             type="number"
@@ -77,7 +75,7 @@ const ExpenseForm = () => {
             onChange={amountChangeHandler}
           />
         </div>
-        <div className="new-expense__controls">
+        <div className="new-expense__control">
           <label>Date</label>
           <input
             type="date"
