@@ -1,10 +1,22 @@
-import './NewExpense.css';
-import ExpenseForm from './ExpenseForm.js';
+import React from "react";
 
-const NewExpense = () => {
-  return <div className="new-expense">
-      <ExpenseForm />
-  </div>;
+import ExpenseForm from "./ExpenseForm";
+import "./NewExpense.css";
+
+const NewExpense = (props) => {
+  const saveExpenseDataHandler = (enteredExpenseData) => {
+    const expenseData = {
+      ...enteredExpenseData,
+      id: Math.random().toString(),
+    };
+    props.onAddExpense(expenseData);
+  };
+
+  return (
+    <div className="new-expense">
+      <ExpenseForm onSaveExpenseData={saveExpenseDataHandler} />
+    </div>
+  );
 };
 
 export default NewExpense;
