@@ -1,6 +1,11 @@
+import React from "react";
+
 import "./Card.css";
 
-export default function Card(props) {
-  const s = "card " + props.className;
-  return <div className={s}>{props.children}</div>;
-}
+const Card = (props) => {
+  const classes = "card " + props.className;
+
+  return <div className={classes}>{props.children}</div>;
+};
+
+export default Card;
